create modules folder and refactored code
