+++ index.html
@@ -9,7 +9,7 @@
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/4.1.1/animate.min.css">
 </head>
 
-<body onload=displayTime();>
+<body>
     <header>
         <nav id="navbar">
             <a href="">Awesome Books</a>
@@ -66,7 +66,7 @@
     <footer>
        <p>  Copyright ... </p>
     </footer>
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
 </body>
 
 </html>
