+++ modules/book.js
@@ -1,0 +1,11 @@
+export default class Book {
+  // Book Class has a title and author
+  title;
+
+  author;
+
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
+  }
+}
